+++ platformio/include/config.h
@@ -65,7 +65,7 @@
 
     // Weekend cell background (will be dithered)
     #define COLOR_CALENDAR_WEEKEND_BG      GxEPD_YELLOW
-    #define DITHER_CALENDAR_WEEKEND        50  // 50% dithering for weekend cells
+    #define DITHER_CALENDAR_WEEKEND        25  // 25% dithering for weekend cells
 
     // Previous/next month days dithering
     #define DITHER_CALENDAR_OUTSIDE_MONTH  20  // 20% dithering for prev/next month days